feat: better 404 page and generate-utils merged with build-utils
